ci: improve workflows by ensuring jobs only run when needed
