+++ .github/workflows/ci.yml
@@ -5,19 +5,50 @@ on:
       - main
     paths:
       - "src/**"
-      - "Cargo.*"
+      - "tests/**"
+      - "Cargo.lock"
   pull_request:
     branches:
       - main
     paths:
       - "src/**"
-      - "Cargo.*"
+      - "tests/**"
+      - "Cargo.lock"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 permissions:
   contents: read
+  pull-requests: read
 jobs:
+  changes:
+    runs-on: ubuntu-24.04
+    permissions:
+      pull-requests: read
+    outputs:
+      src: ${{ steps.filter.outputs.src }}
+      tests: ${{ steps.filter.outputs.tests }}
+    steps:
+      - name: Checkout
+        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5
+
+      - name: Check for changes
+        uses: dorny/paths-filter@de90cc6fb38fc0963ad72b210f1f284cd68cea36 # v3
+        id: filter
+        with:
+          filters: |
+            src:
+              - 'src/**'
+              - 'Cargo.lock'
+            tests:
+              - 'tests/**'
+
   rustfmt:
     runs-on: ubuntu-24.04
-    if: github.actor != 'renovate[bot]'
+    needs: changes
+    if: github.actor != 'renovate[bot]' && needs.changes.outputs.src == 'true'
     name: rustfmt / stable
     steps:
       - name: Checkout
@@ -41,6 +72,8 @@ jobs:
 
   clippy:
     runs-on: ubuntu-24.04
+    needs: changes
+    if: needs.changes.outputs.src == 'true' || needs.changes.outputs.tests == 'true'
     name: clippy / stable
     steps:
       - name: Checkout repository
@@ -59,34 +92,18 @@ jobs:
           skipPush: true
           useDaemon: false
 
+      - name: Cache Rust
+        uses: Swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2
+        with:
+          shared-key: "clippy"
+
       - name: Cargo Clippy
         run: nix develop -c cargo clippy -- -Dwarnings
 
-  typos:
-    runs-on: ubuntu-24.04
-    if: github.actor != 'renovate[bot]'
-    steps:
-      - name: Checkout
-        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5
-
-      - name: Install Nix
-        uses: cachix/install-nix-action@7ab6e7fd29da88e74b1e314a4ae9ac6b5cda3801 # v31
-        with:
-          github_access_token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Setup Cachix
-        uses: cachix/cachix-action@0fc020193b5a1fa3ac4575aa3a7d3aa6a35435ad # v16
-        with:
-          name: purpleclay
-          authToken: "${{ secrets.GH_CACHIX }}"
-          skipPush: true
-          useDaemon: false
-
-      - name: Run Typos
-        run: nix develop -c typos
-
   test:
     runs-on: ${{ matrix.os }}
+    needs: changes
+    if: needs.changes.outputs.src == 'true' || needs.changes.outputs.tests == 'true'
     name: test / ${{ matrix.os }}
     strategy:
       matrix:
@@ -108,11 +125,46 @@ jobs:
           skipPush: true
           useDaemon: false
 
+      - name: Cache Rust
+        uses: Swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2
+        with:
+          shared-key: "test-${{ matrix.os }}"
+
       - name: Run tests
         run: nix develop -c cargo test
 
+  build-fast:
+    runs-on: ubuntu-24.04
+    needs: changes
+    if: github.event_name == 'pull_request' && needs.changes.outputs.src == 'true'
+    name: build-fast / linux_x86_64_musl
+    steps:
+      - name: Checkout
+        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5
+
+      - name: Install stable
+        uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: x86_64-unknown-linux-musl
+
+      - name: Setup Zig
+        uses: goto-bus-stop/setup-zig@abea47f85e598557f500fa1fd2ab7464fcb39406 # v2.2.1
+        with:
+          version: 0.13.0
+
+      - name: Install cargo-zigbuild
+        run: cargo install cargo-zigbuild
+
+      - name: Cache
+        uses: Swatinem/rust-cache@98c8021b550208e191a6a3145459bfc9fb29c4c0 # v2
+
+      - name: Build with zigbuild
+        run: cargo zigbuild --locked --release --target x86_64-unknown-linux-musl
+
   build:
     runs-on: ${{ matrix.platform.runs-on }}
+    needs: changes
+    if: github.ref == 'refs/heads/main' && needs.changes.outputs.src == 'true'
     name: build / ${{ matrix.platform.name }}
     strategy:
       matrix:
