+++ .github/workflows/code-security.yml
@@ -5,15 +5,20 @@ on:
       - main
     paths:
       - "src/**"
-      - "Cargo.*"
+      - "Cargo.lock"
   pull_request:
     branches:
       - main
     paths:
       - "src/**"
-      - "Cargo.*"
+      - "Cargo.lock"
   schedule:
     - cron: "0 5 * * *"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 jobs:
   trivy-scan:
     name: code scan
