+++ .github/workflows/nix.yml
@@ -7,6 +7,7 @@ on:
       - "*.nix"
       - "**/*.nix"
       - "Cargo.lock"
+      - "flake.lock"
   pull_request:
     branches:
       - main
@@ -14,6 +15,12 @@ on:
       - "*.nix"
       - "**/*.nix"
       - "Cargo.lock"
+      - "flake.lock"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 permissions:
   contents: write
 jobs:
