+++ .github/workflows/shellcheck.yml
@@ -10,6 +10,11 @@ on:
       - main
     paths:
       - "scripts/**"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
 permissions:
   contents: read
 jobs:
