+++ .gitignore
@@ -18,6 +18,7 @@
 !Cargo.toml
 !Cargo.lock
 !src/*.rs
+!tests/*.rs
 !build.rs
 !benches/*.rs
 !benches/fixtures/**/*.yaml
